scraper/requirements.txt
selenium
schedule
webdriver-manager
pytesseract
Pillow
pytest
